add missing modification

+++ JavaExercise/src/test/java/com/sky/divideConquer/MatrixComputeFibonacciTest.java
@@ -7,11 +7,11 @@ import org.testng.annotations.Test;
 public class MatrixComputeFibonacciTest extends CommonTest<Integer, Integer> {
     @Test
     public void test() {
-        check(10, null);
-        check(100, null);
-        check(1000, null);
-        check(10000, null);
-        check(100000, null);
+        check(10, (Integer) null);
+        check(100, (Integer)null);
+        check(1000, (Integer)null);
+        check(10000, (Integer)null);
+        check(100000, (Integer)null);
     }
 
     @Override
