+++ JavaExercise/src/test/java/com/sky/exercise/UglyNumberTest.java
@@ -8,10 +8,10 @@ public class UglyNumberTest extends CommonTest<Integer, Integer> {
 
     @Test
     public void test() {
-        check(3, null);
-        check(10, null);
-        check(99, null);
-        check(911, null);
+        check(3, (Integer) null);
+        check(10, (Integer) null);
+        check(99, (Integer) null);
+        check(911, (Integer) null);
     }
 
     @Override
